requirements.txt
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
pyautogui>=0.9.54
customtkinter>=5.2.0
numpy>=1.26.0